requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchvision==0.16.0+cpu
torchaudio==2.1.0+cpu
ultralytics==8.2.103
pdf2image
pillow
pytesseract
streamlit
numpy
